Fix: Don't tag words twice if a (sub-)paragraph is seen multiple times during paragraph breaking.
Since the last commit, words that are part of a node lists that is fed
to pre_linebreak_filter more than once are tagged multiple times.
Therefore, the same word might appear more than once in text output.

This commit implements a quick hack to fix that.  Before tagging a word
in a node list, the nodes preceeding start node and following end node
of the current word are checked for tag nodes with identical semantics.
If they exist and represent the same word, no new tag is inserted.
This is only poor-man's solution, because in the case of a mis-match,
two different word tags are inserted into the node list.

A better alternative were to validate and update tags, so that node
tags represent a state as late as possible before paragraph breaking.
Yet another alternative is not to recurse into node lists, so that the
nodes representing a word are never seen twice.